+++ src/FsPickler.CSharp/Properties/AssemblyInfo.cs
@@ -4,10 +4,10 @@ using System.Reflection;
 [assembly: AssemblyTitleAttribute("FsPickler")]
 [assembly: AssemblyProductAttribute("FsPickler")]
 [assembly: AssemblyCopyrightAttribute("© Eirik Tsarpalis.")]
-[assembly: AssemblyVersionAttribute("1.3.1")]
-[assembly: AssemblyFileVersionAttribute("1.3.1")]
+[assembly: AssemblyVersionAttribute("1.3.7")]
+[assembly: AssemblyFileVersionAttribute("1.3.7")]
 namespace System {
     internal static class AssemblyVersionInformation {
-        internal const string Version = "1.3.1";
+        internal const string Version = "1.3.7";
     }
 }