+++ src/FsPickler.Json/AssemblyInfo.fs
@@ -4,9 +4,9 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("FsPickler")>]
 [<assembly: AssemblyProductAttribute("FsPickler")>]
 [<assembly: AssemblyCopyrightAttribute("© Eirik Tsarpalis.")>]
-[<assembly: AssemblyVersionAttribute("1.3.1")>]
-[<assembly: AssemblyFileVersionAttribute("1.3.1")>]
+[<assembly: AssemblyVersionAttribute("1.3.7")>]
+[<assembly: AssemblyFileVersionAttribute("1.3.7")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.3.1"
+    let [<Literal>] Version = "1.3.7"